Change RLS interface to RLS.with
The `RLS.with` method takes a block and keyword arguments `rls_disabled`,
`user`, and `tenant`.

Within a `with` block, the corresponding database settings are set
either via `SET LOCAL` (if currently within a transaction) or `SET
SESSION`, and reverted after leaving the block. Any additional database
connections created within the block inherit the current settings.

Settings are *NOT* propagated across threads, since there doesn't seem
to be a safe way to accomplish this. The previous approach of setting a
global instance variable on the RLS module is prone to race conditions,
so it has been removed. It is the caller's responsibility to ensure that
any new threads call `RLS.with` themselves to establish the tenancy
scope.

TODO:
* Allow the set of arguments (currently `user` and `tenant`) to be
  defined dynamically via configuration
* Allow RLS to be configured separately for multiple databases, e.g.
  some databases may have different variables for tenancy, or may not
  have RLS available at all.