Isolate RLS state across threads when connection is shared
With `use_transactional_tests` in Rails 5.1+, the database connection is
shared among threads. Since the RLS state is specific to a thread (so
that in integration tests, for example, the test code may be operating
with RLS disabled while the server under test is operating with a user
or tenant set), the state needs to be restored in each thread as it
gains control of the connection.

This is implemented by wrapping all of the database adapter methods with
`steal_thread_for_rls!`, which applies the thread-local RLS state if the
connection was last used by a different thread.